solo falta favoritos

+++ MoviesApp/Entities/DTO/DetailDTO/MovieDetailDTO.swift
@@ -20,7 +20,7 @@ struct MovieDetailDTO: Decodable {
     let original_language: String?
     let original_title: String?
     let overview: String?
-    let popularity: Int?
+    let popularity: Double?
     let poster_path: String?
     let production_companies: [CompaniesDTO]?
     let production_countries: [ProductionCountriesDTO]?
